requirements.txt
faster-whisper
google-generativeai>=0.8.3
openai
opencv-python
pyaudio
ultralytics
pydantic
numpy